+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Copy token button in UI.
 - Context analyzer tool.
 - New false positive utterances.
+- Links to platform in entitycore tool outputs.
 
 ### Fixed
 - Make sure .env is not read in tests.